# src/main/resources/static/openapi/inventario-openapi-specs.yaml
openapi: 3.1.0
info:
  title: Inventario OpenAPI definition
  description: "An API for a web-based system to manage inventory stocks and handle stocks analytics"
  version: 1.0.0
  contact:
    email: pshameed99@gmail.com
servers:
  - url: http://localhost:8080/inventario/api/v1
...
paths:
...
  /users:
    get:
      tags:
        - UserEntity
      summary: Retrieve all users
      security:
        - BearerAuth: [ ]
      responses:
        200:
          description: List of users
        403:
          description: Forbidden (Insufficient permissions)
    post:
      tags:
        - UserEntity
      summary: Create a new user
      security:
        - BearerAuth: [ ]
      responses:
        201:
          description: User created successfully
        403:
          description: Forbidden (Insufficient permissions)
  /users/changePassword:
    patch:
      tags:
        - UserEntity
      summary: Change password for the authenticated user
      security:
        - BearerAuth: [ ]
      requestBody:
        description: Password change request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        201:
          description: Password updated successfully
        403:
          description: Forbidden (Insufficient permissions)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PasswordChangeRequest:
      type: object
      properties:
        newPassword:
          type: string
      required:
        - newPassword
security:
  - BearerAuth: []